ADDED: EnhancedInvoice

+++ src/main/java/com/bridgelabz/cabinvoicegenerator/Ride.java
@@ -1,10 +1,10 @@
 package com.bridgelabz.cabinvoicegenerator;
 
 public class Ride {
-	public int distance;
+	public double distance;
     public int time;
 
-    public Ride(int distance, int time) {
+    public Ride(double distance, int time) {
         this.distance = distance;
         this.time = time;
     }
